# perf: html5 attribute parsing