… gumbo-parser/src/hashmap.c
… nokogiri.gemspec