@flavorjones 2024-12-27
@stevecheckoway I would really love your feedback on this when you get a few minutes. ❤️