@flavorjones 2024-12-27
@stevecheckoway Thanks for the review. Yes, I was sloppy with malloc failures to get something up and running; I'll fix.

I was super curious about what you think of introducing the hashmap into libgumbo? Just want to make sure you don't think it's a silly thing to do.